./go: Update to v1.5.0, download on demand
With the updates to the `./go` script from the upstream `go-template`
file, it's no longer necessary to have `mbland/go-script-bash` attached
to this repository as a submodule.

After pulling in this commit, run `rm -rf script/go-script-bash` to
download the new version of `mbland/go-script-bash` upon the next
invocation of the `./go` script.