hubot/sgi.js: Add Robot.react to Hubot >= v3.0.0
Addresses the issue described in #15. In short, hubot-test-helper v1.8.0
depends on hubot-v3.0.0, both are rewrites from CoffeeScript to ES2015,
and consequently hubot-slack adds Robot.react to a CoffeeScript-
compatible wrapper function, leading to smoke and integration test
failures.

The implementation of slackapi/hubot-slack#429 (CoffeeScript-to-ES2015
rewrite) should include a fix that will eliminate the need for this
workaround.